% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_harbours.R
\name{add_harbours}
\alias{add_harbours}
\title{Add_harbours}
\usage{
add_harbours(x, y)
}
\arguments{
\item{x}{gps dataset}

\item{y}{harbour polygons}
}
\value{
A gps datset with an extra column "SI_HARB"
}
\description{
A function that will add a column "SI_HARB" to the gps datset
SI_HARB = 1 is where the vessel is considered in harbour
}
