% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAcceleration.R
\name{CalcAcceleration}
\alias{CalcAcceleration}
\title{Calculate Acceleration in meters.s-2}
\usage{
CalcAcceleration(
  trip.path,
  col.time = "time_stamp",
  col.speed = "speed",
  speed.units = "knots"
)
}
\arguments{
\item{trip.path}{xxx}

\item{col.time}{Varible name containing time (default "time_stamp")}

\item{col.speed}{Variable name containing speed (default "speed")}

\item{speed.units}{Speed units (default "knots"), any other interpeted as ms}
}
\value{
A track with additional variable named "acceleration"
}
\description{
Calculate Acceleration in meters.s-2
}
