% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDist.R
\name{CalcDist}
\alias{CalcDist}
\title{Calculate distance between consequitve points}
\usage{
CalcDist(trip.path, col.time = "time_stamp")
}
\arguments{
\item{trip.path}{xxx}

\item{col.time}{Varible name containing time (default "time_stamp")}
}
\value{
The original sf object with a column "DISTANCE.nm" in nautic miles
}
\description{
Calculate distance between positions using great circle distance
calculations see ?sf::st_distance
}
