% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDistBetweenNearestNeighbours.R
\name{CalcDistBetweenNearestNeighbours}
\alias{CalcDistBetweenNearestNeighbours}
\title{Title}
\usage{
CalcDistBetweenNearestNeighbours(trip.path, nnn = 9)
}
\arguments{
\item{trip.path}{xxx}

\item{nnn}{Number of points to calculate (default 9)}
}
\value{
The original sf object with n columns showing the distances in meters to the n nearest neighbours
}
\description{
Calculate Distance between the point and its n nearest neighbours.
Use of the method implemented in RANN::nn2 which must be the fastest existing
in R for this purpose
}
