% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcHeading.R
\name{CalcHeading}
\alias{CalcHeading}
\title{Calculate Heading}
\usage{
CalcHeading(trip.path)
}
\arguments{
\item{trip.path}{xxx}
}
\value{
The original sf object with a column "HEADING.deg" in degrees
}
\description{
Calculate Heading
}
