% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSpeed.R
\name{CalcSpeed}
\alias{CalcSpeed}
\title{Calculate speed between two positions}
\usage{
CalcSpeed(trip.path, col.time = "time_stamp", verbose = FALSE)
}
\arguments{
\item{trip.path}{xxx}

\item{col.time}{Varible name containing time (default "time_stamp")}

\item{verbose}{Details of output (default FALSE)}
}
\value{
The original sf object with a column "DIFFTIME.secs" in seconds,
"SPEED.kn" in knots and eventually "DISTANCE.nm" in nautic miles if
CalcDist hasn't been applied before.
}
\description{
Calculate speed between two positions
}
