% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcStraigthness.R
\name{CalcStraigthness}
\alias{CalcStraigthness}
\title{Calculate Straightness}
\usage{
CalcStraigthness(trip.path, col.Dir)
}
\arguments{
\item{trip.path}{xxx}

\item{col.Dir}{geometry column??}
}
\value{
The original sf object with a column "abs.HeadingChange"
}
\description{
Absolute difference between heading at t+1 and heading at t.
Should be problematic for values around 0-360 => over-estimation of the
heading change!
}
