% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_trips_pol.R
\name{define_trips_pol}
\alias{define_trips_pol}
\title{define_trips_pol}
\usage{
define_trips_pol(
  x,
  min_dur = 0.5,
  max_dur = 72,
  split_trips = T,
  preserve_all = F
)
}
\arguments{
\item{x}{gps dataset}

\item{min_dur}{the minimum trip length (hours)}

\item{max_dur}{the maximum trip length (hours)}

\item{split_trips}{If the trip is longer than the maximum hours, it will try to split
the trip into two or more trips, if there is long enough intervals between pings}

\item{preserve_all}{Should all pings inside harbour be preserved also?}
}
\value{
a gps datset
}
\description{
Use the columns "SI_HARB" to determine when a vessel is on a trip. A trip is defined
from when it leaves the harbour till it returns
}
