% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfTOsf.R
\name{dfTOsf}
\alias{dfTOsf}
\title{dfTOsf}
\usage{
dfTOsf(df, type = "multipoint", epsg = NA)
}
\arguments{
\item{df}{data.frame or matrix to be converted in sf multipoints object.}

\item{type}{DEFAULT "multipoint", type of sf object, "linestring" for line, "polygon" for polygon.}

\item{epsg}{either a valid "crs" object or a valid numeric espg code}
}
\value{
a sf multipoints objects
}
\description{
\code{dfTOsf} Spatialize coordinates.
}
\examples{

obj <- matrix(c( 181109.5, 182616.6, 333729.7, 331218.0), ncol = 2)

obj.pol <- matrix( c(181414.4, 333221.7,  181473.8, 333122.7, 179791.8, 330095.1,  179732.4, 330194.1), byrow =TRUE, ncol =2) 

sf.pts <- dfTOsf(obj)
sf.lin <- dfTOsf(obj, type = "linestring")
sf.pol <- dfTOsf(obj.pol, type = "polygon")

plot( sf::st_geometry(sf.pol), axes = TRUE, ylim = c(330095.1, 334000))
plot( sf::st_geometry(sf.lin), add = TRUE, col ="red")
plot( sf::st_geometry(sf.pts), add = TRUE, col ="blue")  

}
\author{
Julien Rodriguez, \email{julien.rodriguez@limagrain.com}
}
