% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_ais_timestep.R
\name{downsample_ais_timestep}
\alias{downsample_ais_timestep}
\title{downsample_ais_timestep}
\usage{
downsample_ais_timestep(x, unit = "min", length = 1)
}
\arguments{
\item{x}{gps dataset}

\item{unit}{sec/min/hour - the unit of the timestep}

\item{length}{number of units between timestep}
}
\description{
downsample_ais_timestep
}
