% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harbours}
\alias{harbours}
\title{Polygons of 942 scandinavian harbours}
\format{
A data frame with 942 rows and 2 variables:
\describe{
\item{SI_HARB}{harbour id (all values 1)}
\item{geometry}{polygons}
...
}
}
\source{
\url{http://....}
}
\usage{
harbours
}
\description{
A dataset containing ..
}
\keyword{datasets}
