% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_ais.R
\name{interpolate_ais}
\alias{interpolate_ais}
\title{Interpolate AIS}
\usage{
interpolate_ais(
  x,
  headingAdjustment = 0,
  min_gap = 2,
  max_gap = 120,
  source = "AIS"
)
}
\arguments{
\item{x}{A gps dataset in a standard format}

\item{headingAdjustment}{numeric. If using heading instead of course}

\item{min_gap}{a numeric value. Gaps more than this should be interpolated (minutes)}

\item{max_gap}{a numeric value. Gaps less than this should be interpolated (minutes)}

\item{source}{character. The type of the gps dataset}
}
\value{
a gps dataset that has added the interpolated points
}
\description{
Sometimes there is gaps in the gps dataset, this is a way
to fill out these gaps. Inspired very much from
vmstools::interpolateTacsat
}
