% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramb_functions.R
\name{rb_arcdist}
\alias{rb_arcdist}
\title{rb_acdist}
\usage{
rb_arcdist(lat, lon, lat1 = NULL, lon1 = NULL, scale = "nmi")
}
\arguments{
\item{lat}{Latitude of first coordinate or list with lat, lon of first coordinate.}

\item{lon}{Longitude of first coordinate or list with lat, lon of second coordinate.}

\item{lat1}{If lat and lon are vectors of lat,lon positions, then lat1 and lon1 must be given as the second set of positions.}

\item{lon1}{If lat and lon are vectors of lat,lon positions, then lat1 and lon1 must be given as the second set of positions.}

\item{scale}{"nmi" (default) returns value in nautical miles, any other value in kilometers}
}
\value{
A single vector of distances between pairs of points
}
\description{
Computes distances between lat/lon data points.
}
