% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramb_functions.R
\name{rb_define_trips}
\alias{rb_define_trips}
\title{rb_define_trips}
\usage{
rb_define_trips(vid = vid, time = time, hid = hid)
}
\arguments{
\item{vid}{vector containing vessel id}

\item{hid}{vector containing harbour id, value assumed NA or zero if out of harbour}
}
\value{
A vector containing unique trip id, negative if in harbour, positive if 
out of harbour
}
\description{
rb_define_trips
}
