% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{track01}
\alias{track01}
\title{AIS tracks of some scandinavian vessel}
\format{
A data frame with 942 rows and 2 variables:
\describe{
\item{vessel_id}{xxx}
\item{time_stamp}{xxx}
\item{lon}{xxx}
\item{lat}{xxx}
\item{speed}{xxx}
\item{course}{xxx}
\item{gear}{xxx}
\item{behaviour}{xxx}
...
}
}
\source{
\url{http://....}
}
\usage{
track01
}
\description{
A dataset containing ..
}
\keyword{datasets}
