% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_RF.R
\name{tune_RF}
\alias{tune_RF}
\title{tune_RF}
\usage{
tune_RF(formula, data, nb.param = 5, ntree.try = 50)
}
\arguments{
\item{formula}{a model formula built using stats::as.formula}

\item{data}{a data.frame, used to train the model}

\item{nb.param}{DEFAULT = 5, number of hyperparameters tested, two designs of nb.param X nb.param are tested}

\item{ntree.try}{DEFAULT = 50, Number of trees to grow in each random-forest used to test the combination of hyperparameters.}
}
\value{
a list containing optimal values for hyperparameters min.node.size and mtry
}
\description{
\code{tune_RF} Automatic optimization of hyperparameters used in a random forest model (mtry and min.node.size). Two designs of nb.param X nb.param are tested using Out-of-Bag error estimate. The full design is then predicted from a random-forest model.
}
\examples{

library(sp)
require(ranger)
data(meuse)    
set.seed(100)

tune_RF( zinc ~ dist + soil + ffreq, meuse )
tune_RF( copper ~ dist + soil + ffreq, meuse )
optim.rf <- tune_RF( lead ~ dist + soil + ffreq, meuse )

mod.rf <- ranger(lead ~ dist + soil + ffreq, meuse , 
importance = "impurity", mtry =  optim.rf$mtry, min.node.size =  optim.rf$min.node.size, num.trees = 500, write.forest = TRUE)

}
\author{
Julien Rodriguez, \email{julien.rodriguez@ifremer.com}
}
