% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WKSSFGEO-package.R
\docType{package}
\name{WKSSFGEO-package}
\alias{WKSSFGEO}
\alias{WKSSFGEO-package}
\title{WKSSFGEO: Track functions for AIS/VMS data}
\description{
Some convenient functions developed as WKSSFGEO meeting for vessel tracks
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
